NavBarCreateChar moved to folder container and added to NavBar folder
