+++ src/components/NavBar/NavBarCreateChar/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import './NavBar/NavBar.css';
+import '../NavBar.css';
 
 const NavBarCreateChar = ({ totalWeight }) => {
   const characterWeight = Math.round((totalWeight) * 1000) / 1000;
